Create HR Simulator
Built an HR system that can be used to track a manager's approval or rejection of an employee's requests to go on vacation, leave or overtime. I built a class that can be used to add the names of Managers and the names of employees and I built a logic to help regulate what each manager can do for an employee. An employee cannot approve or reject anything for a manager. The system helps regulate that only managers can approve or reject vacations, holidays or overtime for employees.